+++ tests/test_config.py
@@ -3,7 +3,6 @@ import os
 import pytest
 
 from jupytext.config import (
-    JupytextConfiguration,
     find_jupytext_configuration_file,
     load_jupytext_configuration_file,
 )
@@ -57,36 +56,113 @@ def test_load_jupytext_configuration_file(tmpdir, config_file):
 
     if config_file.endswith(("jupytext", ".toml")):
         full_config_path.write(
-            """default_jupytext_formats = "ipynb,py:percent"
-default_notebook_metadata_filter = "all"
-default_cell_metadata_filter = "all"
+            """formats = "ipynb,py:percent"
+notebook_metadata_filter = "all"
+cell_metadata_filter = "all"
 """
         )
     elif config_file.endswith(".yml"):
         full_config_path.write(
-            """default_jupytext_formats: ipynb,py:percent
-default_notebook_metadata_filter: all
-default_cell_metadata_filter: all
+            """formats: ipynb,py:percent
+notebook_metadata_filter: all
+cell_metadata_filter: all
 """
         )
     elif config_file.endswith(".json"):
         full_config_path.write(
-            """{"default_jupytext_formats": "ipynb,py:percent",
-"default_notebook_metadata_filter": "all",
-"default_cell_metadata_filter": "all"
+            """{"formats": "ipynb,py:percent",
+"notebook_metadata_filter": "all",
+"cell_metadata_filter": "all"
 }
 """
         )
     elif config_file.endswith(".py"):
         full_config_path.write(
-            """c.default_jupytext_formats = "ipynb,py:percent"
-c.default_notebook_metadata_filter = "all"
-c.default_cell_metadata_filter = "all"
+            """c.formats = "ipynb,py:percent"
+c.notebook_metadata_filter = "all"
+c.cell_metadata_filter = "all"
 """
         )
 
     config = load_jupytext_configuration_file(str(full_config_path))
-    config = JupytextConfiguration(**config)
-    assert config.default_jupytext_formats == "ipynb,py:percent"
-    assert config.default_notebook_metadata_filter == "all"
-    assert config.default_cell_metadata_filter == "all"
+    assert config.formats == "ipynb,py:percent"
+    assert config.notebook_metadata_filter == "all"
+    assert config.cell_metadata_filter == "all"
+
+
+@pytest.mark.parametrize(
+    "content_toml,formats_short_form",
+    [
+        (
+            """# always pair ipynb notebooks to py:percent files
+formats = "ipynb,py:percent"
+""",
+            "ipynb,py:percent",
+        ),
+        (
+            """# always pair ipynb notebooks to py:percent files
+formats = ['ipynb', 'py:percent']
+""",
+            "ipynb,py:percent",
+        ),
+        (
+            """# always pair ipynb notebooks to py:percent files
+formats = ["ipynb", "py:percent"]
+""",
+            "ipynb,py:percent",
+        ),
+        (
+            """# always pair ipynb notebooks to py:percent files
+formats = "ipynb,py:percent"
+""",
+            "ipynb,py:percent",
+        ),
+        (
+            """# Pair notebooks in subfolders of 'notebooks' to scripts in subfolders of 'scripts'
+[formats]
+"notebooks/" = "ipynb"
+"scripts/" = "py:percent"
+""",
+            "notebooks///ipynb,scripts///py:percent",
+        ),
+        (
+            """# Pair notebooks in subfolders of 'notebooks' to scripts in subfolders of 'scripts'
+[formats]
+"notebooks" = "ipynb"
+"scripts" = "py:percent"
+""",
+            "notebooks///ipynb,scripts///py:percent",
+        ),
+    ],
+)
+def test_jupytext_formats(tmpdir, content_toml, formats_short_form):
+    jupytext_toml = tmpdir.join("jupytext.toml")
+    jupytext_toml.write(content_toml)
+
+    config = load_jupytext_configuration_file(str(jupytext_toml))
+    assert config.formats == formats_short_form
+
+
+def test_deprecated_formats_cause_warning(
+    tmpdir, content_toml="default_jupytext_formats = 'ipynb,md'"
+):
+    jupytext_toml = tmpdir.join("jupytext.toml")
+    jupytext_toml.write(content_toml)
+
+    config = load_jupytext_configuration_file(str(jupytext_toml))
+    with pytest.warns(FutureWarning, match="use 'formats'"):
+        assert config.default_formats(str(tmpdir.join("test.md"))) == "ipynb,md"
+
+
+@pytest.mark.parametrize(
+    "option_name",
+    ["notebook_metadata_filter", "cell_metadata_filter", "cell_markers"],
+)
+def test_deprecated_options_cause_warning(tmpdir, option_name):
+    jupytext_toml = tmpdir.join("jupytext.toml")
+    jupytext_toml.write(f"default_{option_name} = 'value'")
+    config = load_jupytext_configuration_file(str(jupytext_toml))
+    fmt = {}
+    with pytest.warns(FutureWarning, match=f"use '{option_name}'"):
+        config.set_default_format_options(fmt)
+        assert fmt[option_name] == "value"