+++ docs/CHANGELOG.md
@@ -1,6 +1,17 @@
 Jupytext ChangeLog
 ==================
 
+1.11.0 (2021-03-18)
+-------------------
+
+**Fixed**
+- The `jupytext.toml` config file can now be used together with the `jupytext` pre-commit hook ([#752](https://github.com/mwouts/jupytext/issues/752))
+- The `notebook_extensions` option of the `jupytext.toml` file now works ([#746](https://github.com/mwouts/jupytext/issues/746))
+
+**Changed**
+- The options in `jupytext.toml` where renamed to match the `jupytext` metadata in the text notebooks. One should now use `formats` rather than `default_jupytext_formats` and `notebook_metadata_filter` rather than `default_notebook_metadata_filter` ([#753](https://github.com/mwouts/jupytext/issues/753))
+
+
 1.10.3 (2021-03-07)
 -------------------